requirements.txt
langchain
langchain-community
openai
chromadb
pypdf
tiktoken
faiss-cpu
python-dotenv
